[Console] merged Command and Console components #386
