+++ Installer/EntityFinder.php
@@ -39,7 +39,8 @@ class EntityFinder
         '/Mock/',
         '/TestUnit/',
         '/Exception/',
-        '/Commands/',
+        '/Console/',
+        '/Command/',
         '/Installer/',
         '/Assets/',
         '/Renderer/',
